requirements.txt
google-genai
openai

python-dotenv
schedule

Pillow
moviepy
imageio-ffmpeg

instagrapi
requests
pydantic
tenacity
